+++ crates/starknet-devnet-server/src/api/json_rpc/mod.rs
@@ -51,7 +51,7 @@ use crate::api::json_rpc::models::{
     BroadcastedInvokeTransactionEnumWrapper, SimulateTransactionsInput,
 };
 use crate::api::serde_helpers::{empty_params, optional_params};
-use crate::restrictive_methods::is_json_rpc_method_restricted;
+use crate::restrictive_mode::is_json_rpc_method_restricted;
 use crate::rpc_core::error::RpcError;
 use crate::rpc_core::request::RpcMethodCall;
 use crate::rpc_core::response::{ResponseResult, RpcResponse};
