+++ crates/starknet-devnet-server/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod api;
 mod config;
 pub mod error;
-pub mod restrictive_methods;
+pub mod restrictive_mode;
 pub mod rpc_core;
 /// handlers for axum server
 pub mod rpc_handler;
